[feat] 플레이어와 아이템 상호작용
* 플레이어 컨트롤러 주석정리 포함
* 물풍선 파워 int로 변경
* 아이템 상호작용을 위한 주석 해제

+++ Assets/Develop/SHW/Scripts/PlayerStatus.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 public class PlayerStatus : MonoBehaviour
 {
     [SerializeField] public float speed;       // ÇÃ·¹ÀÌ¾î ¼Óµµ
-    [SerializeField] public float power;       // ÆøÅºÆÄ¿ö
+    [SerializeField] public int power;         // ÆøÅºÆÄ¿ö
     [SerializeField] public int bombCount;     // ÆøÅº¼ö 
 
-    public bool isBubble;
+    public bool isBubble;                      // ¹°¹æ¿ï¿¡ °¤Èù »óÅÂ
 }
